chore: moved easter egg files in weather-easter-egg folder
