+++ src/components/content/heroes/aurora/weather-easter-egg/precipitation-effect.tsx
