+++ src/components/content/heroes/aurora/weather-easter-egg/sun-percentage-calculator.ts
