+++ src/components/content/heroes/aurora/weather-easter-egg/sun.test.ts
